+++ packages/create-zenith/templates/css/zenith.config.js
@@ -1,5 +1,4 @@
 export default {
     pagesDir: 'src/pages',
-    router: false,
     typescriptDefault: true
 };